Unit behaviour improved; Sprites changes
The 3 units can be directly controlled, sent to 3 specific locations, sent to steal gold from the foe's base, and told to mine from your own gold mine. A basic game loop and victory condition, as well as inter-player interactivity, will be planned out. This interactivity will involve spending gold on something useful (perhaps 'buildings' that can be placed on the board, or simple, branching unit upgrades such as movement speed or gold carrying capacity), as well as interactivity with the other team's units in some way (perhaps by 'fighting' in some sense). Then the code will be refactored to be more generic so that behaviour can be applied to the right team's units.

Sprites and the main background have been changed.