Change of tack for now
For now, I want to take what I have built so far and turn it into a single player game that is between Warcraft 3 and Plants v Zombies. The goal (for now) will be to survive as many turns as possible. Certain functionality to this end has been implemented, with more to come.