Further Small Tweaks
Further changes. Nearly at a point of having a MVP that shows what I had in mind. Next steps might include new/revised sprites, and a simple pause screen, as well as a start screen that is returned to when the player runs out of lives and which displays both the current high score and the controls.