Incremental Update
Extra functionality implemented, for example the ability to gather stone as well as gold. The stone will be a cost for building certain towers. Appearance tweaked, among other things. Need to significantly tidy up the code in various areas. Need to add a simple title screen that explains the game and controls, displays a high score, and is returned to after a Game Over.